+++ src/auth/configs/salesforce.config.ts
@@ -1,16 +1,10 @@
 import { registerAs } from '@nestjs/config';
 import { ISalesforceConfig } from './interfaces/salesforce-config.interface';
-import { NODE_ENV } from '../../environments';
 
 export default registerAs('salesforce', (): ISalesforceConfig => {
-  const baseURL =
-    NODE_ENV === 'production'
-      ? 'https://login.salesforce.com/services/oauth2'
-      : 'https://test.salesforce.com/services/oauth2';
-
-  const authorizationURL = `${baseURL}/authorize`; // `${process.env.SALESFORCE_URL}/services/oauth2/authorize`
-  const tokenURL = `${baseURL}/token`;
-  const profileURL = `${baseURL}/userinfo`;
+  const authorizationURL = `${process.env.SALESFORCE_URL}/services/oauth2/authorize`;
+  const tokenURL = `${process.env.SALESFORCE_URL}/services/oauth2/token`;
+  const profileURL = `${process.env.SALESFORCE_URL}/services/oauth2/userinfo`;
 
   return {
     clientID: process.env.OAUTH2_SALESFORCE_CLIENT_ID,