+++ src/common/middlewares/tenant.middleware.ts
@@ -17,7 +17,6 @@ export class TenantMiddleware implements NestMiddleware {
         req.headers.host
       ).toString();
 
-      // console.log('source', source);
       const tenantName: string = source
         .split('.')[0]
         .toString()
@@ -25,7 +24,6 @@ export class TenantMiddleware implements NestMiddleware {
         .replace('undefined', '')
         .replace(/\:\d{1,4}/, '');
 
-      // console.log('TenantMiddleware/SetTenant', tenantName);
       set('tenant', tenantName);
 
       next();