+++ src/oem/main/oem-rules/oem-discount-rules/oem-discount-rule.dto/oem-discount-rule.dto.ts
@@ -53,7 +53,7 @@ export class DiscountRuleDto {
   @Transform(({ value }) => {
     return String(value)?.trim();
   })
-  @Validate(IsDiscountRuleExists)
+  //@Validate(IsDiscountRuleExists)
   discountRuleName: string;
 
   /**