+++ src/oem/main/oem-rules/oem-shading-rules/oem-shading-rule.dto/oem-shading-rule.dto.ts
@@ -60,7 +60,7 @@ export class ShadingRuleDto {
   @Transform(({ value }) => {
     return String(value)?.trim();
   })
-  @Validate(IsShadingRuleExists)
+  //@Validate(IsShadingRuleExists)
   shadingRuleName: string;
 
   /**